chore(*): removing ssr-test and making demos-dev ssr
