+++ projects/igniteui-angular/src/lib/tabs/tabs/tabs.component.spec.ts
@@ -45,7 +45,6 @@ describe('IgxTabs', () => {
         ];
 
         TestBed.configureTestingModule({
-            declarations: [],
             imports: [
                 NoopAnimationsModule,
                 RouterTestingModule.withRoutes(testRoutes),
