--- projects/ssr-test/src/app/app.component.html
@@ -1,7 +1,0 @@
-<igx-navbar></igx-navbar>
-
-<igx-list>
-    <igx-list-item></igx-list-item>
-</igx-list>
-
-<router-outlet></router-outlet>
