+++ src/app/app.component.ts
@@ -1,12 +1,30 @@
 import { Component, OnInit, ViewChild, HostBinding } from '@angular/core';
-import { Router, NavigationStart, NavigationEnd } from '@angular/router';
+import { Router, NavigationStart, NavigationEnd, RouterLinkActive, RouterLink, RouterOutlet } from '@angular/router';
 import { filter } from 'rxjs/operators';
 import { IgxNavigationDrawerComponent, IgxIconService } from 'igniteui-angular';
+import { PageHeaderComponent } from './pageHeading/pageHeading.component';
+import { IgxIconComponent } from '../../projects/igniteui-angular/src/lib/icon/icon.component';
+import { NgFor, NgIf } from '@angular/common';
+import { IgxNavDrawerTemplateDirective, IgxNavDrawerItemDirective, IgxNavDrawerMiniTemplateDirective } from '../../projects/igniteui-angular/src/lib/navigation-drawer/navigation-drawer.directives';
 
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss']
+    styleUrls: ['./app.component.scss'],
+    standalone: true,
+    imports: [
+        IgxNavigationDrawerComponent,
+        IgxNavDrawerTemplateDirective,
+        IgxNavDrawerItemDirective,
+        NgFor,
+        RouterLinkActive,
+        RouterLink,
+        IgxIconComponent,
+        NgIf,
+        IgxNavDrawerMiniTemplateDirective,
+        PageHeaderComponent,
+        RouterOutlet
+    ]
 })
 export class AppComponent implements OnInit {
     @HostBinding('attr.id')
