+++ src/tsconfig.app.json
@@ -3,15 +3,18 @@
     "compilerOptions": {
         "outDir": "../out-tsc/app",
         "types": [
-            "node"
-        ]
+            "node",
+            "hammerjs"
+        ],
+        "allowSyntheticDefaultImports": true
     },
-  "files": [
-    "main.ts",
-    "polyfills.ts",
-    "../projects/igniteui-angular-i18n/src/i18n/expand-type.ts"
-  ],
-  "include": [
-    "**/*.d.ts"
-  ]
+    "files": [
+        "main.ts",
+        "main.server.ts",
+        "server.ts",
+        "../projects/igniteui-angular-i18n/src/i18n/expand-type.ts"
+    ],
+    "include": [
+        "**/*.d.ts"
+    ]
 }
